moved schrodinet to qmcflow and removed anything superfluous for flows
