--- schrodinet/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Documentation about pyCHAMP"""
-
-from .__version__ import __version__
-
-__author__ = "Nicolas Renaud"
-__email__ = 'n.renaud@esciencecenter.nl'
